refactor(WithUsageDataPreviewPanel): Extract UsageData to its own file
