+++ src/WingmanDataTrail/MetricVizPanel/WithUsageDataPreviewPanel.tsx
@@ -1,5 +1,3 @@
-import { css, cx } from '@emotion/css';
-import { type GrafanaTheme2 } from '@grafana/data';
 import {
   SceneCSSGridLayout,
   sceneGraph,
@@ -7,7 +5,7 @@ import {
   type SceneComponentProps,
   type SceneObjectState,
 } from '@grafana/scenes';
-import { Button, Dropdown, Icon, Menu, Tooltip, useStyles2, type IconName } from '@grafana/ui';
+import { type IconName } from '@grafana/ui';
 import React from 'react';
 
 import { logger } from 'tracking/logger/logger';
@@ -26,6 +24,8 @@ import {
   METRICS_VIZ_PANEL_HEIGHT_WITH_USAGE_DATA_PREVIEW,
   type MetricVizPanel,
 } from 'WingmanDataTrail/MetricVizPanel/MetricVizPanel';
+
+import { UsageData, type UsageSectionProps } from './UsageData';
 
 type SortBy = Exclude<SortingOption, 'related'>;
 
@@ -158,135 +158,5 @@ export class WithUsageDataPreviewPanel extends SceneObjectBase<WithUsageDataPrev
         />
       </div>
     );
-  };
-}
-// TODO: pass full MetricUsageDetailsObject into UsageSectionProps?
-interface UsageSectionProps {
-  usageType: MetricUsageType;
-  usageCount: number;
-  singularUsageType: string;
-  pluralUsageType: string;
-  icon: IconName;
-  usageDetails: MetricUsageDetails;
-}
-
-function UsageData({
-  usageType,
-  usageCount,
-  singularUsageType,
-  pluralUsageType,
-  icon,
-  usageDetails,
-}: Readonly<UsageSectionProps>) {
-  const styles = useStyles2(getStyles);
-
-  let dashboardItems: Array<{ label: string; value: string; count: number }> = [];
-  if (usageDetails.usageType === 'dashboard-usage') {
-    const { dashboards } = usageDetails;
-    dashboardItems = Object.entries(dashboards)
-      .map(([label, dashboardInfo]) => ({
-        label,
-        value: `/d/${dashboardInfo.uid}`,
-        count: dashboardInfo.count,
-      }))
-      .sort((a, b) => b.count - a.count);
-  }
-
-  return (
-    <div className={styles.usageContainer} data-testid="usage-data-panel">
-      {usageDetails.usageType === 'dashboard-usage' ? (
-        <>
-          <Dropdown
-            placement="right-start"
-            overlay={
-              <Menu style={{ maxWidth: '240px', maxHeight: '245px', overflowY: 'auto' }}>
-                {dashboardItems.map((item) => (
-                  <Menu.Item
-                    key={item.value}
-                    label=""
-                    url={item.value}
-                    target="_blank"
-                    className={styles.menuItem}
-                    component={() => (
-                      <Tooltip
-                        content={`Found ${item.count} ${item.count === 1 ? 'time' : 'times'} in ${item.label}`}
-                        placement="right"
-                      >
-                        <div className={styles.menuItemContent}>
-                          <Icon name="external-link-alt" /> {item.label} ({item.count})
-                        </div>
-                      </Tooltip>
-                    )}
-                  />
-                ))}
-              </Menu>
-            }
-          >
-            <Button
-              variant="secondary"
-              size="sm"
-              tooltip={`Metric used in ${usageCount} dashboards. Click to view them.`}
-              className={cx(styles.usageItem, styles.clickableUsageItem)}
-            >
-              <span data-testid={usageType}>
-                <Icon name={icon} style={{ marginRight: '4px' }} /> {usageCount}
-              </span>
-            </Button>
-          </Dropdown>
-        </>
-      ) : (
-        <Tooltip
-          content={`Metric is used in ${usageCount} ${usageCount === 1 ? singularUsageType : pluralUsageType}`}
-          placement="top"
-        >
-          <span className={styles.usageItem} data-testid={usageType}>
-            <Icon name={icon} /> {usageCount}
-          </span>
-        </Tooltip>
-      )}
-    </div>
-  );
-}
-
-function getStyles(theme: GrafanaTheme2) {
-  return {
-    usageContainer: css({
-      display: 'flex',
-      flexDirection: 'row',
-      justifyContent: 'flex-start',
-      gap: '17px',
-      padding: '8px 12px',
-      border: `1px solid ${theme.colors.border.weak}`,
-      borderTopWidth: 0,
-      backgroundColor: theme.colors.background.primary,
-      alignItems: 'center',
-    }),
-    usageItem: css({
-      display: 'flex',
-      alignItems: 'center',
-      gap: '4px',
-      color: theme.colors.text.secondary,
-      opacity: '65%',
-    }),
-    clickableUsageItem: css({
-      backgroundColor: 'transparent',
-      border: 'none',
-    }),
-    menuItem: css({
-      color: theme.colors.text.primary,
-      textDecoration: 'none',
-      '&:hover': {
-        color: theme.colors.text.link,
-      },
-    }),
-    menuItemContent: css({
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      whiteSpace: 'nowrap',
-      color: theme.colors.text.primary,
-      '&:hover': {
-        color: theme.colors.text.link,
-      },
-    }),
   };
 }
